requirements.txt
traci
sumolib
python-dotenv
pyproj
paho-mqtt